Write controller for handling POST from genre submission form
First, it validates and sanitises the data using express-validator's
validator.body function

Then, it extracts any errors resulting from this with the
validator.validationResult function

A new Genre object is created using the sanitised data

If there are
any errors, the submission form is rerendered with error messages and
the sanitised data

Otherwise, we check if the chosen genre already exists.

If it does, we
simply redirect the user to its existing page.

If it doesn't, we save the new genre (in the database) and redirect
the user to the new genre page